+++ packages/lexical-headless/src/index.ts
@@ -6,268 +6,7 @@
  *
  */
 
-import type {
-  DOMChildConversion,
-  DOMConversion,
-  DOMConversionFn,
-  EditorState,
-  EditorThemeClasses,
-  GridSelection,
-  LexicalEditor,
-  LexicalNode,
-  NodeSelection,
-  RangeSelection,
-} from 'lexical';
+import {createHeadlessEditor} from './LexicalHeadlessEditor';
+import {$generateHtmlFromNodes, $generateNodesFromDOM} from './LexicalHtml';
 
-import {
-  $cloneWithProperties,
-  $sliceSelectedTextNodeContent,
-} from '@lexical/selection';
-import {$getRoot, $isElementNode, $isTextNode, createEditor} from 'lexical';
-import {Class} from 'utility-types';
-
-export function createHeadlessEditor(editorConfig?: {
-  disableEvents?: boolean;
-  editorState?: EditorState;
-  namespace?: string;
-  nodes?: ReadonlyArray<Class<LexicalNode>>;
-  onError: (error: Error) => void;
-  parentEditor?: LexicalEditor;
-  readOnly?: boolean;
-  theme?: EditorThemeClasses;
-}): LexicalEditor {
-  const editor = createEditor(editorConfig);
-  editor._headless = true;
-
-  [
-    'registerDecoratorListener',
-    'registerRootListener',
-    'registerMutationListeners',
-    'getRootElement',
-    'setRootElement',
-    'getElementByKey',
-    'focus',
-    'blur',
-  ].forEach((method) => {
-    editor[method] = () => {
-      throw new Error(`${method} is not supported in headless mode`);
-    };
-  });
-
-  return editor;
-}
-
-export function $appendNodesToHTML(
-  editor: LexicalEditor,
-  selection: RangeSelection | NodeSelection | GridSelection | null,
-  currentNode: LexicalNode,
-  parentElement: HTMLElement | DocumentFragment,
-): boolean {
-  let shouldInclude = selection != null ? currentNode.isSelected() : true;
-  const shouldExclude =
-    $isElementNode(currentNode) && currentNode.excludeFromCopy('html');
-  let clone = $cloneWithProperties<LexicalNode>(currentNode);
-  clone =
-    $isTextNode(clone) && selection != null
-      ? $sliceSelectedTextNodeContent(selection, clone)
-      : clone;
-  const children = $isElementNode(clone) ? clone.getChildren() : [];
-  const {element, after} = clone.exportDOM(editor);
-
-  if (!element) {
-    return false;
-  }
-
-  const fragment = new DocumentFragment();
-
-  for (let i = 0; i < children.length; i++) {
-    const childNode = children[i];
-    const shouldIncludeChild = $appendNodesToHTML(
-      editor,
-      selection,
-      childNode,
-      fragment,
-    );
-
-    if (
-      !shouldInclude &&
-      $isElementNode(currentNode) &&
-      shouldIncludeChild &&
-      currentNode.extractWithChild(childNode, selection, 'html')
-    ) {
-      shouldInclude = true;
-    }
-  }
-
-  if (shouldInclude && !shouldExclude) {
-    element.append(fragment);
-    parentElement.append(element);
-
-    if (after) {
-      const newElement = after.call(clone, element);
-      if (newElement) element.replaceWith(newElement);
-    }
-  } else {
-    parentElement.append(fragment);
-  }
-
-  return shouldInclude;
-}
-
-export function $generateHtmlFromNodes(
-  editor: LexicalEditor,
-  selection?: RangeSelection | NodeSelection | GridSelection | null,
-): string {
-  if (document == null || window == null) {
-    throw new Error(
-      'To use $generateHtmlFromNodes in headless mode please initialize a headless browser implementation such as JSDom before calling this function.',
-    );
-  }
-
-  const container = document.createElement('div');
-  const root = $getRoot();
-  const topLevelChildren = root.getChildren();
-
-  for (let i = 0; i < topLevelChildren.length; i++) {
-    const topLevelNode = topLevelChildren[i];
-    $appendNodesToHTML(editor, selection, topLevelNode, container);
-  }
-
-  return container.innerHTML;
-}
-
-function getConversionFunction(
-  domNode: Node,
-  editor: LexicalEditor,
-): DOMConversionFn | null {
-  const {nodeName} = domNode;
-
-  const cachedConversions = editor._htmlConversions.get(nodeName.toLowerCase());
-
-  let currentConversion: DOMConversion | null = null;
-
-  if (cachedConversions !== undefined) {
-    cachedConversions.forEach((cachedConversion) => {
-      const domConversion = cachedConversion(domNode);
-
-      if (domConversion !== null) {
-        if (
-          currentConversion === null ||
-          currentConversion.priority < domConversion.priority
-        ) {
-          currentConversion = domConversion;
-        }
-      }
-    });
-  }
-
-  return currentConversion !== null ? currentConversion.conversion : null;
-}
-
-const IGNORE_TAGS = new Set(['STYLE']);
-
-function $createNodesFromDOM(
-  node: Node,
-  editor: LexicalEditor,
-  forChildMap: Map<string, DOMChildConversion> = new Map(),
-  parentLexicalNode?: LexicalNode | null | undefined,
-): Array<LexicalNode> {
-  let lexicalNodes: Array<LexicalNode> = [];
-
-  if (IGNORE_TAGS.has(node.nodeName)) {
-    return lexicalNodes;
-  }
-
-  let currentLexicalNode = null;
-  const transformFunction = getConversionFunction(node, editor);
-  const transformOutput = transformFunction ? transformFunction(node) : null;
-  let postTransform = null;
-
-  if (transformOutput !== null) {
-    postTransform = transformOutput.after;
-    currentLexicalNode = transformOutput.node;
-
-    if (currentLexicalNode !== null) {
-      for (const [, forChildFunction] of forChildMap) {
-        currentLexicalNode = forChildFunction(
-          currentLexicalNode,
-          parentLexicalNode,
-        );
-
-        if (!currentLexicalNode) {
-          break;
-        }
-      }
-
-      if (currentLexicalNode) {
-        lexicalNodes.push(currentLexicalNode);
-      }
-    }
-
-    if (transformOutput.forChild != null) {
-      forChildMap.set(node.nodeName, transformOutput.forChild);
-    }
-  }
-
-  // If the DOM node doesn't have a transformer, we don't know what
-  // to do with it but we still need to process any childNodes.
-  const children = node.childNodes;
-  let childLexicalNodes = [];
-
-  for (let i = 0; i < children.length; i++) {
-    childLexicalNodes.push(
-      ...$createNodesFromDOM(
-        children[i],
-        editor,
-        forChildMap,
-        currentLexicalNode,
-      ),
-    );
-  }
-
-  if (postTransform != null) {
-    childLexicalNodes = postTransform(childLexicalNodes);
-  }
-
-  if (currentLexicalNode == null) {
-    // If it hasn't been converted to a LexicalNode, we hoist its children
-    // up to the same level as it.
-    lexicalNodes = lexicalNodes.concat(childLexicalNodes);
-  } else {
-    if ($isElementNode(currentLexicalNode)) {
-      // If the current node is a ElementNode after conversion,
-      // we can append all the children to it.
-      currentLexicalNode.append(...childLexicalNodes);
-    }
-  }
-
-  return lexicalNodes;
-}
-
-/**
- * How you parse your html string to get a document is left up to you. In the browser you can use the native
- * DOMParser API to generate a document (see clipboard.ts), but to use in a headless environment you can use JSDom
- * or an equivilant library and pass in the document here.
- */
-export function $generateNodesFromDOM(
-  dom: Document,
-  editor: LexicalEditor,
-): Array<LexicalNode> {
-  let lexicalNodes = [];
-  const elements: Array<Node> = dom.body ? Array.from(dom.body.childNodes) : [];
-  const elementsLength = elements.length;
-
-  for (let i = 0; i < elementsLength; i++) {
-    const element = elements[i];
-
-    if (!IGNORE_TAGS.has(element.nodeName)) {
-      const lexicalNode = $createNodesFromDOM(element, editor);
-
-      if (lexicalNode !== null) {
-        lexicalNodes = lexicalNodes.concat(lexicalNode);
-      }
-    }
-  }
-
-  return lexicalNodes;
-}
+export {$generateHtmlFromNodes, $generateNodesFromDOM, createHeadlessEditor};
